lvlm_models/pyproject.toml
[project]
name = "lvlm_models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "datasets>=3.5.1",
    "openai>=1.79.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
